+++ lnd_methods/unauthenticated/subscribe_to_wallet_status.d.ts
@@ -6,13 +6,16 @@ import {
 /** The wallet has yet to be created */
 export type SubscribeToWalletStatusAbsentEvent = EmptyObject;
 
-/** The wallet is activated and ready for all requests */
+/** The wallet is activated and has started working */
 export type SubscribeToWalletStatusActiveEvent = EmptyObject;
 
 export type SubscribeToWalletStatusErrorEvent = Error;
 
 /** The wallet is inactive because it is locked */
 export type SubscribeToWalletStatusLockedEvent = EmptyObject;
+
+/** The wallet is ready for all RPC requests */
+export type SubscribeToWalletStatusReadyEvent = EmptyObject;
 
 /** The wallet is in the process of starting */
 export type SubscribeToWalletStatusStartingEvent = EmptyObject;