+++ lnd_responses/rpc_wallet_state_as_state.js
@@ -1,6 +1,7 @@
 const stateAbsent = 'NON_EXISTING';
 const stateActive = 'RPC_ACTIVE';
 const stateLocked = 'LOCKED';
+const stateReady = 'SERVER_ACTIVE';
 const stateStarting = 'UNLOCKED';
 const stateWaiting = 'WAITING_TO_START';
 
@@ -15,6 +16,7 @@ const stateWaiting = 'WAITING_TO_START';
     [is_absent]: <Wallet Not Created Bool>
     [is_active]: <Wallet Is Active Bool>
     [is_locked]: <Wallet File Encrypted And Wallet Not Active Bool>
+    [is_ready]: <Wallet Is Ready For All RPC Calls Bool>
     [is_starting]: <Wallet Is Starting Up Bool>
     [is_waiting]: <Wallet Is Waiting To Start Bool>
   }
@@ -37,6 +39,9 @@ module.exports = args => {
 
   case stateLocked:
     return {is_locked: true};
+
+  case stateReady:
+    return {is_active: true, is_ready: true};
 
   case stateStarting:
     return {is_starting: true};