+++ test/lnd_methods/offchain/test_delete_failed_pay_attempts.js
@@ -9,6 +9,24 @@ const tests = [
     error: [400, 'ExpectedAuthenticatedLndToDeleteFailedAttempts'],
   },
   {
+    args: {id: 1},
+    description: 'A payment id is expected to be a hash',
+    error: [400, 'ExpectedPaymentHashToDeleteFailedPayAttempts'],
+  },
+  {
+    args: {
+      id: Buffer.alloc(32).toString('hex'),
+      lnd: {
+        default: {
+          deleteAllPayments: ({}, cbk) => cbk(),
+          deletePayment: ({}, cbk) => cbk({details: 'unknown'}),
+        },
+      },
+    },
+    description: 'An unsupported error is returned',
+    error: [501, 'DeleteFailedPaymentAttemptsMethodNotSupported'],
+  },
+  {
     args: {lnd: {default: {deleteAllPayments: ({}, cbk) => cbk('err')}}},
     description: 'An unexpected error is returned',
     error: [503, 'UnexpectedErrorDeletingFailedAttempts', {err: 'err'}],